Change scenario, correct spelling and add glossary entries
Scince we changed the features of our web application, searching for
station names is no longe available. Therefore scenario 2 must be
changed slightly. Now the user searches for an address or determins his
position with the gps feature.
In addition I corrected some spelling mistakes and added glossary
entries for the words "Feature" and "Station".